feat: unwrap Pty namespace to flat exports + barrel
Uses --name service to avoid collision with existing pty.ts type defs.
Also adds --name flag to the unwrap-namespace script.
